Smarter result counts
This implementation comes from discussions with @wschella and @cecemel

The previous solution counted the amount of distinct objects which may
not represent the amount of triples starting with that subject.

An initial assumption would be that the following query would yield the
desired result-set.

    SELECT (COUNT (DISTINCT *) AS ?count)
    WHERE {
      <http://example.com/42> ?p ?o.
    }

That does not seem to be the case in the currently used sparql
implementation.

This implementation is an elegant solution to the problem taking the
form of a subselect query.

    SELECT (COUNT *) AS ?count)
    WHERE {
      SELECT DISTINCT ?p ?o
      WHERE {
        <http://example.com/42> ?p ?o.
      }
    }

The inner query constructs a result set with one match for each `?p ?o`
tuple.  Counting these (already distinct) rows thus counts all triples
which which have that subject.

The inverse implementation is analogous.